organize code and get rid of cond statements
